Refactor: Split and namespace 2FA model
SecondartyAuthentication model was was a wrapper for 2
responsibilities:
- Container for secondary authentication operations and their time
  validity.
- Direct OTP secondary authentication logic.

This refactor splits both responsibilities into their respective
classes, namespaced under SecondaryAuthentication module.
