+++ cosmetics-web/app/controllers/responsible_persons/delete_notification_controller.rb
@@ -15,7 +15,7 @@ class ResponsiblePersons::DeleteNotificationController < SubmitApplicationContro
 private
 
   def current_operation
-    SecondaryAuthentication::DELETE_NOTIFICATION
+    SecondaryAuthentication::Operations::DELETE_NOTIFICATION
   end
 
   def set_responsible_person
