+++ cosmetics-web/app/controllers/responsible_persons/team_members_controller.rb
@@ -109,6 +109,6 @@ private
 
   # See: SecondaryAuthenticationConcern
   def current_operation
-    SecondaryAuthentication::INVITE_USER
+    SecondaryAuthentication::Operations::INVITE_USER
   end
 end
