+++ cosmetics-web/spec/features/account/change_mobile_number_spec.rb
@@ -19,7 +19,7 @@ RSpec.describe "Changing mobile number", :with_2fa, :with_stubbed_mailer, :with_
     click_on "Your account"
     expect_to_be_on_my_account_page
 
-    wait_for = SecondaryAuthentication::TIMEOUTS[SecondaryAuthentication::CHANGE_MOBILE_NUMBER]
+    wait_for = SecondaryAuthentication::Operations::TIMEOUTS[SecondaryAuthentication::Operations::CHANGE_MOBILE_NUMBER]
     travel_to((wait_for + 1).seconds.from_now)
 
     click_on "Change mobile number"
