+++ cosmetics-web/spec/features/account/change_password_spec.rb
@@ -13,7 +13,7 @@ RSpec.describe "Changing password", :with_2fa, :with_stubbed_mailer, :with_stubb
       click_on "Your account"
       expect_to_be_on_my_account_page
 
-      wait_for = SecondaryAuthentication::TIMEOUTS[SecondaryAuthentication::CHANGE_PASSWORD]
+      wait_for = SecondaryAuthentication::Operations::TIMEOUTS[SecondaryAuthentication::Operations::CHANGE_PASSWORD]
       travel_to((wait_for + 1).seconds.from_now)
 
       click_on "Change password"
